Only generate a filename
Using mkstemp won't work on Windows, where the file permissions are locked down
such that nothing else can write to it. All we need is a unique filename, not
a file created for us.

The node id is unique for a given tree, so just use that; because we use a
unique temporary directory per tree, there is no need for further complication
here.